Hold Ctrl+Shift+Alt to temporarily suppress the full redness
Hold Ctrl+Shift+Alt (Ctrl+Shift+Option on macOS) to temporarily suppress
the full redness. This was requested by 1P2D on the Blender Extensions
website:

Would be good to add a feature that allows me to temporarily see through
the red screen, before saving. It happened to me to have that red screen
after not using Blender for a few minutes, and to forget where I was at
with my model - CTRL+S completely blindly can be dangerous sometimes.

https://extensions.blender.org/add-ons/you-are-autosave/reviews/#review-1483